Minimap camera bug fix; can recover when mainCamera exits the scene. Also checks for ground boundingBox and picks a suitable orthographicSize to display everything.
There was a bug that made impossible to operate with the minimap when the mainCamera wasn't looking at the ground. That's been solved and code in the Update function has been optimized.

Minimap Camera has been disabled on startup cause it becomes active in Start function. That way it doesn't show in the editor at first necessarily.

Now it can check for "Ground" GO and takes its boundingBox so it adapts the camera orthographicSize and, that way, displays the whole scenario.